Renamed this package to client instead of http
