--- http/middleware.go
@@ -1,1 +1,0 @@
-package http
